JS: Added Last Input Variable
This variable will keep track of the previous input entered. The goal is
to rehash the input each time it is changed. The reason the onchange
event is not being used for this is that it does not change until the
user clicks off of the form.

Instead a timer will be used to call a function that will check if the
input has changed, and if it has only then will it hash again. This
requires the lastInput variable to store the last accepted input and use
it for comparisons.