JS: Fixed Scoping issues for variable input in hashInput()
Previously there was a definition for input at the top of the function
that was intended to be used for all lower scopes in the function.

This proved to be impossible due to it being out of the scope for the
WASM code.

Instead, one is now located within the WASM related code and one for the
JS code.